backend/apps/records/filters.py
"""Django-filter FilterSet for FinancialRecord."""

import django_filters
from .models import FinancialRecord


class FinancialRecordFilter(django_filters.FilterSet):
    date_from = django_filters.DateFilter(field_name="date", lookup_expr="gte")
    date_to   = django_filters.DateFilter(field_name="date", lookup_expr="lte")
    min_amount = django_filters.NumberFilter(field_name="amount", lookup_expr="gte")
    max_amount = django_filters.NumberFilter(field_name="amount", lookup_expr="lte")

    class Meta:
        model   = FinancialRecord
        fields  = ["type", "category", "date_from", "date_to", "min_amount", "max_amount"]
